service API completed

+++ src/main/java/ru/slisarenko/kalita/dto/WalletDTO.java
@@ -3,5 +3,5 @@ package ru.slisarenko.kalita.dto;
 import lombok.Builder;
 
 @Builder
-public record WalletDTO(String id, String operationType, Long amount) {
+public record WalletDTO(String id, String operationType, Double amount) {
 }
